+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,12 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.5] - Unreleased
+
+### Changed
+
+- With `portable` enabled, a long option (for example, `--physical` for the `cd` built-in) passed to a built-in that defines one is now rejected as an error, instead of being accepted as an extension.
+
 ## [3.3.4] - 2026-07-31
 
 ### Changed
@@ -441,6 +447,7 @@ later.
 
 - Initial release of the shell
 
+[3.3.5]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.5
 [3.3.4]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.4
 [3.3.3]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.3
 [3.3.2]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.3.2